+++ projects/demo/src/app/app.module.ts
@@ -2,11 +2,8 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { AppComponent } from './app.component';
 import { AngularCesiumModule, AngularCesiumWidgetsModule } from 'angular-cesium';
-import { HttpClientModule } from '@angular/common/http';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { FormsModule } from '@angular/forms';
-import { TracksDialogComponent } from './components/tracks-layer/track-dialog/track-dialog.component';
-import { TracksLayerComponent } from './components/tracks-layer/tracks-layer.component';
 import { LabelLayerExampleComponent } from './components/label-layer-example/label-layer-example.component';
 import { EllipseLayerExampleComponent } from './components/ellipse-layer-example/ellipse-layer-example.component';
 import { CircleLayerExampleComponent } from './components/circle-layer-example/circle-layer-example.component';
@@ -24,7 +21,6 @@ import { ModelsLayerExampleComponent } from './components/models-layer/models-la
 import { PolygonPerformanceTestComponent } from './components/polygon-layer/polygon-performance-test/polygon-performance-test.component';
 import { HippodromeLayerExampleComponent } from './components/hippodrome-layer/hippodrome-layer-example.component';
 import { KeyboardControlLayerComponent } from './components/keyboard-control-layer/keyboard-control-layer.component';
-import { LayerOrderComponent } from './components/layer-order/layer-order.component';
 import { ContextMenuLayerComponent } from './components/context-menu-layer/context-menu-layer.component';
 import { SelectionLayerExampleComponent } from './components/selection-layer/selection-layer-example.component';
 import { PolygonsEditorExampleComponent } from './components/editor-layer/polygons-editor-layer/polygons-editor-example.component';
@@ -40,15 +36,6 @@ import { TrackEntityLayerComponent } from './components/track-entity-layer/track
 import { CzmlLayerExampleComponent } from './components/czml-layer/czml-layer-example.component';
 import { EllipsesEditorExampleComponent } from './components/editor-layer/ellipses-editor-layer/ellipses-editor-example.component';
 import { AppMaterialModule } from './app.material.module';
-import { InMemoryCache } from 'apollo-cache-inmemory';
-import { environment } from '../environments/environment';
-import { HttpLink, HttpLinkModule } from 'apollo-angular-link-http';
-import { Apollo, APOLLO_OPTIONS, ApolloModule } from 'apollo-angular';
-import { MaxValidatorDirective } from './layout/settings-form/max-validtor.directive';
-import { MinValidatorDirective } from './layout/settings-form/min-validator.directive';
-import { MainNavbarComponent } from './layout/main-navbar/main-navbar.component';
-import { SidenavToolbarComponent } from './layout/sidenav-toolbar/sidenav-toolbar.component';
-import { SettingsFormComponent } from './layout/settings-form/settings-form.component';
 import { SingleEntityOnMapExampleComponent } from './components/single-entity-on-map-example/single-entity-on-map-example.component';
 import { MyCustomContextMenuComponent } from './components/context-menu-layer/context-menu/my-custom-context-menu.component';
 import { BillboardLayerExampleComponent } from './components/billboard-layer-example/billboard-layer-example.component';
@@ -66,13 +53,9 @@ import { WallLayerExampleComponent } from './components/wall-layer-example/wall-
     AngularCesiumWidgetsModule,
     BrowserAnimationsModule,
     AppMaterialModule,
-    HttpClientModule,
-    ApolloModule,
-    HttpLinkModule,
   ],
   declarations: [
     AppComponent,
-    TracksLayerComponent,
     BillboardLayerExampleComponent,
     WallLayerExampleComponent,
     CylinderLayerExampleComponent,
@@ -88,7 +71,6 @@ import { WallLayerExampleComponent } from './components/wall-layer-example/wall-
     MapEventsExampleComponent,
     ArcLayerExampleComponent,
     PointLayerExampleComponent,
-    TracksDialogComponent,
     SingleEntityOnMapExampleComponent,
     DemoMapComponent,
     DemoMultipleMapsComponent,
@@ -97,7 +79,6 @@ import { WallLayerExampleComponent } from './components/wall-layer-example/wall-
     PolygonPerformanceTestComponent,
     KeyboardControlLayerComponent,
     HippodromeLayerExampleComponent,
-    LayerOrderComponent,
     ContextMenuLayerComponent,
     SelectionLayerExampleComponent,
     PolygonsEditorExampleComponent,
@@ -112,24 +93,10 @@ import { WallLayerExampleComponent } from './components/wall-layer-example/wall-
     TrackEntityLayerComponent,
     CzmlLayerExampleComponent,
     EllipsesEditorExampleComponent,
-    MaxValidatorDirective,
-    MinValidatorDirective,
-    MainNavbarComponent,
-    SidenavToolbarComponent,
-    SettingsFormComponent,
     MyCustomContextMenuComponent,
   ],
-  entryComponents: [TracksDialogComponent, MyCustomContextMenuComponent],
+  entryComponents: [MyCustomContextMenuComponent],
   bootstrap: [AppComponent]
 })
 export class AppModule {
-  constructor(apollo: Apollo, httpLink: HttpLink) {
-    apollo.create({
-      link: httpLink.create({
-        uri: environment.server + '/graphql',
-      }),
-      cache: new InMemoryCache(),
-    });
-  }
-
 }