+++ projects/demo/src/app/components/ellipse-layer-example/ellipse-layer-example.component.ts
@@ -1,7 +1,6 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { AcLayerComponent, AcNotification, ActionType } from 'angular-cesium';
 import { Observable } from 'rxjs';
-import { TracksDataProvider } from '../../utils/services/dataProvider/tracksDataProvider.service';
 import { MockDataProviderService } from '../../utils/services/dataProvider/mock-data-provider.service';
 import { map } from 'rxjs/operators';
 
@@ -9,7 +8,6 @@ import { map } from 'rxjs/operators';
   selector: 'ellipse-layer-example',
   templateUrl: 'ellipse-layer-example.component.html',
   styleUrls: ['ellipse-layer-example.component.css'],
-  providers: [TracksDataProvider]
 })
 export class EllipseLayerExampleComponent implements OnInit {
   @ViewChild(AcLayerComponent, {static: false}) layer: AcLayerComponent;